Refactor: call it PDO check
